requirements.txt
flask==3.1.0
gunicorn==23.0.0
cdsapi==0.7.6
xarray==2024.6.0
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
matplotlib==3.9.2
Pillow==10.4.0
scipy==1.13.1
h5netcdf==1.3.0
requests
boto3
